Refactor testing to pass typeguard checks
